% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{add_distinct_values_from_columns}
\alias{add_distinct_values_from_columns}
\title{Add Distinct Values From Columns
This function find sthe number of distinct values from one or more columns
and returns the sum of those.}
\usage{
add_distinct_values_from_columns(tbl, columns)
}
\arguments{
\item{tbl}{A tibble}

\item{columns}{a list of strings that are names of columns in the tibble}
}
\description{
Add Distinct Values From Columns
This function find sthe number of distinct values from one or more columns
and returns the sum of those.
}
