% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_list_columns.R
\name{concatenate_all_list_columns}
\alias{concatenate_all_list_columns}
\title{This function will concatenate all list columns into character columns.}
\usage{
concatenate_all_list_columns(tbl)
}
\arguments{
\item{tbl}{A Tibble}
}
\description{
This function will concatenate all list columns into character columns.
}
