% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_list_columns.R
\name{concatenate_list_columns}
\alias{concatenate_list_columns}
\title{Concatenate List Columns
This function will concatenate list columns into character columns}
\usage{
concatenate_list_columns(tbl, columns)
}
\arguments{
\item{tbl}{A Tibble}

\item{columns}{A list of strings that are names of columns in data to
be concatenated}
}
\description{
Concatenate List Columns
This function will concatenate list columns into character columns
}
