% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_count_column.R
\name{create_count_column_with_config}
\alias{create_count_column_with_config}
\title{Create Plot Count Dataframe with Config
This function is used to create a count column.}
\usage{
create_count_column_with_config(tbl, config)
}
\arguments{
\item{tbl}{A tibble}

\item{config}{A named list. This list must have a named list named
create_count_column. This list must have names "name" and "complete_columns"}
}
\description{
Create Plot Count Dataframe with Config
This function is used to create a count column.
}
