% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_data_focus_plots}
\alias{create_data_focus_plots}
\title{Create Data Focus Plots}
\usage{
create_data_focus_plots(data_list, config)
}
\arguments{
\item{data_list}{A list of tibbles}

\item{config}{A named list that has the "plot" name. The "plot" value
must be a named list with the "x" name, which is a name in each tibble in the
data list, and a "fill" name with is a list of columns, equal to the length
of the data list.}
}
\description{
Create Data Focus Plots
}
\examples{
data_list <- list(
 "Assays" = dplyr::tribble(
  ~Study, ~Assays,
  "s1",   "a1",
  "s1",   "a2"
),
 "Resources" = dplyr::tribble(
  ~Study, ~Resources,
  "s1",   "r1",
  "s1",   "r1"
 )
)

config <- list(
  "plot" = list(
    "x" = "Study",
    "fill" = list(
      "Assay",
      "Resources"
    )
  )
)
}
