% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{create_data_focus_tables}
\alias{create_data_focus_tables}
\title{Create Data Focus Tables
This function creates a list of tables from on input tibble. The list will
have one table per column listed in the fill_columns list. This function is
used in the study_summary module to create the data_focus plots.}
\usage{
create_data_focus_tables(data, x_column, fill_columns)
}
\arguments{
\item{data}{A tibble}

\item{x_column}{A string that is the name of a column in the data}

\item{fill_columns}{A list of strings that are names of columns in data}
}
\description{
Create Data Focus Tables
This function creates a list of tables from on input tibble. The list will
have one table per column listed in the fill_columns list. This function is
used in the study_summary module to create the data_focus plots.
}
