% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_initiative_activity_plot}
\alias{create_initiative_activity_plot}
\title{Create Initiative Activity Plot}
\usage{
create_initiative_activity_plot(data, x, fill, facet)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}

\item{facet}{A list of string that are names of columns in the data}
}
\description{
Create Initiative Activity Plot
}
