% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{create_merged_table_with_config}
\alias{create_merged_table_with_config}
\title{Create Merged Table With Config}
\usage{
create_merged_table_with_config(group_object, config)
}
\arguments{
\item{group_object}{A named list with tables}

\item{config}{A named list. It must have the names "tables", "columns",
and "join_column". The tables value must be names of tables in the group
object. "join_column" must be a column in all tables. "columns" must be a
list of lists, one per table. Each list must be a subset of that tables
column names.}
}
\description{
Create Merged Table With Config
}
