% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth_functions.R
\name{create_oauth_list}
\alias{create_oauth_list}
\title{Create OAuth List}
\usage{
create_oauth_list(
  config_file = "inst/oauth_config.yaml",
  config = NULL,
  scope = "openid view download modify",
  claims_list = get_synapse_claims_list()
)
}
\arguments{
\item{config_file}{A string, the path to a yaml file}

\item{config}{The list must have the following items:
\itemize{
\item "client_id" An integer or a string containing na integer ie. "1"
\item "client_secret" A string
\item "app_url" A string, the shiny app's URL
}}

\item{scope}{a character vector}

\item{claims_list}{A named list}
}
\description{
Create OAuth List
}
