% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth_functions.R
\name{create_oauth_url_script_html}
\alias{create_oauth_url_script_html}
\title{Create OAuth URL Script HTML}
\usage{
create_oauth_url_script_html(oauth_list)
}
\arguments{
\item{oauth_list}{A named list. The list must have the following items:
\itemize{
\item "endpoint" an object of type httr::oauth_endpoint()
\item "app" an object pf type httr::oauth_app()
\item "scope" a character vector
}}
}
\description{
Create OAuth URL Script HTML
}
