% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_count_df.R
\name{create_plot_count_df}
\alias{create_plot_count_df}
\title{Create Plot Count Dataframe
This function is used to create counts for histogram style plots if the
counting has to be done in a way that preserves certain groups that have a
count of 0.}
\usage{
create_plot_count_df(tbl, factor_columns, complete_columns)
}
\arguments{
\item{tbl}{A tibble}

\item{factor_columns}{A list of strings that are columns in the data. This
should be the aesthetic that is intended to be present in the plot even if
it has zero counts such as the x-axis, or possibly a facet.}

\item{complete_columns}{A list of strings that are columns in the data.
This should be all columns except the color/fill columns.}
}
\description{
Create Plot Count Dataframe
This function is used to create counts for histogram style plots if the
counting has to be done in a way that preserves certain groups that have a
count of 0.
}
