% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_count_df.R
\name{create_plot_count_df_with_config}
\alias{create_plot_count_df_with_config}
\title{Create Plot Count Dataframe with Config
This function is used to create counts for histogram style plots if the
counting has to be done in a way that preserves certain groups that have a
count of 0.}
\usage{
create_plot_count_df_with_config(tbl, config)
}
\arguments{
\item{tbl}{A tibble}

\item{config}{A named list. This list must have a named list named
complete_count. This list must have lists named "factor columns" and
"complete_columns".}
}
\description{
Create Plot Count Dataframe with Config
This function is used to create counts for histogram style plots if the
counting has to be done in a way that preserves certain groups that have a
count of 0.
}
