% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{create_plot_dfs_from_count_df}
\alias{create_plot_dfs_from_count_df}
\title{Create Plot Dataframes From Count Dataframe
See create_plot_df_from_count_df for outputs}
\usage{
create_plot_dfs_from_count_df(column_values, count_columns, data)
}
\arguments{
\item{column_values}{A list of values that exists in the "name" column
of the data}

\item{count_columns}{A list of strings that will the new column names}

\item{data}{A tibble}
}
\description{
Create Plot Dataframes From Count Dataframe
See create_plot_df_from_count_df for outputs
}
