% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_publication_status_plot}
\alias{create_publication_status_plot}
\title{Create Publication Status Plot}
\usage{
create_publication_status_plot(data, x, fill)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}
}
\description{
Create Publication Status Plot
}
