% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_resources_generated_plot}
\alias{create_resources_generated_plot}
\title{Create Resources Generated Plot}
\usage{
create_resources_generated_plot(data, x, fill, facet)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}

\item{facet}{A list of string that are names of columns in the data}
}
\description{
Create Resources Generated Plot
}
