% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_study_timeline_plot}
\alias{create_study_timeline_plot}
\title{Create Study Timeline Plot}
\usage{
create_study_timeline_plot(data, x, fill, facet, y_axis_text = NULL)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}

\item{facet}{A list of string that are names of columns in the data}

\item{y_axis_text}{A list of parameters to
axis.text.y = ggplot2::element_text()}
}
\description{
Create Study Timeline Plot
}
