% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_timeline_module.R
\name{file_upload_timeline_module_server}
\alias{file_upload_timeline_module_server}
\title{File Upload Timeline Server}
\usage{
file_upload_timeline_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A shiny::reactive() that returns a named list. The list must
contain a list named "tables".}

\item{config}{A shiny::reactive() that returns a named list.}
}
\description{
File Upload Timeline Server
}
