% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{filter_list_column}
\alias{filter_list_column}
\title{Filter List Column
This function filters a list column in the input data. Rows are kept if all
items in the values are in the supplied column.}
\usage{
filter_list_column(data, column, values)
}
\arguments{
\item{data}{A Tibble}

\item{column}{A string, that is the name of a column in the data}

\item{values}{A list of values}
}
\description{
Filter List Column
This function filters a list column in the input data. Rows are kept if all
items in the values are in the supplied column.
}
