% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{format_date_columns}
\alias{format_date_columns}
\title{Format Date Columns
This function adds date, year and month columns if the the input has a
createdOn column from Synapse.
If the the input has year or month columns, those are converted to factors.}
\usage{
format_date_columns(data)
}
\arguments{
\item{data}{A tibble}
}
\description{
Format Date Columns
This function adds date, year and month columns if the the input has a
createdOn column from Synapse.
If the the input has year or month columns, those are converted to factors.
}
