% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_plot_data.R
\name{format_plot_data_with_config}
\alias{format_plot_data_with_config}
\title{Format Plot Data With Config}
\usage{
format_plot_data_with_config(tbl, config)
}
\arguments{
\item{tbl}{A Tibble

data <- dplyr::tribble(
~consortium, ~year, ~month,
NA,          2000L, NA,
"c1",        20001, "January"
)}

\item{config}{A named list.

The list must contained an item named "columns" that has an entry
for each column needed in the tibble. Each column must have a "name",
and "type" field. Optional fields include "replace_values", "display_name",
"na_replace", and "default_replace".

The list may also contain an item named drop_na. If the value of drop_na
is True, all rows with na values will be droped.

The list may also contain an item named unnest_cols. All columns in this
list will be unnest with tidyr::unnest()

config <- list(
"columns" = list(
list(
"name" = "consortium",
"display_name" = "Consortium",
"na_replace" = "Not Applicable",
"type" = "character"
),
)
)}
}
\description{
This function runs a tibble through the main data cleanup functions before
being displayed in a plot or data table. See those functions for more
details.
}
