% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{get_distinct_value_from_column}
\alias{get_distinct_value_from_column}
\title{Get Number Of Distinct Values From Column
This function returns the number of distinct values, including NA's in
a column}
\usage{
get_distinct_value_from_column(tbl, column)
}
\arguments{
\item{tbl}{A Tibble}

\item{column}{A string that is the name of a column in the tibble}
}
\description{
Get Number Of Distinct Values From Column
This function returns the number of distinct values, including NA's in
a column
}
