% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth_functions.R
\name{get_extended_synapse_claims_list}
\alias{get_extended_synapse_claims_list}
\title{Get Extended Synapse Claims List}
\usage{
get_extended_synapse_claims_list()
}
\description{
Get Extended Synapse Claims List
}
