% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_column_values.R
\name{get_n_most_common_values}
\alias{get_n_most_common_values}
\title{Get N Most Common Values}
\usage{
get_n_most_common_values(tbl, column, n = 5)
}
\arguments{
\item{tbl}{A tibble}

\item{column}{A string, that is a column in the tbl}

\item{n}{an integer}
}
\description{
Get N Most Common Values
}
