% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{get_synapse_tbl}
\alias{get_synapse_tbl}
\title{Get Synapse Table}
\usage{
get_synapse_tbl(syn, table_id, columns = NULL, limit = NULL)
}
\arguments{
\item{syn}{An object of reticulate::import("synapseclient")$Synapse()}

\item{table_id}{A synapse table id}

\item{columns}{A vector of columns to get from the syanpse table}

\item{limit}{An integer, limits the number of rows in the table}
}
\description{
Get Synapse Table
}
