% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiative_activity_module.R
\name{initiative_activity_module_ui}
\alias{initiative_activity_module_ui}
\alias{initiative_activity_module_server}
\title{initiative_activity_module_server and
initiative_activity_module_server_ui}
\usage{
initiative_activity_module_ui(id)

initiative_activity_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A named list. The list must contain a list named "tables".}

\item{config}{A named list.}
}
\description{
A shiny Module.
}
\keyword{internal}
