% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_studies_module.R
\name{merge_studies_module_ui}
\alias{merge_studies_module_ui}
\alias{merge_studies_module_server}
\title{merge_studies_module_ui and merge_studies_module_server}
\usage{
merge_studies_module_ui(id)

merge_studies_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
