% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_reporting_module.R
\name{milestone_reporting_module_ui}
\alias{milestone_reporting_module_ui}
\alias{milestone_reporting_module_server}
\title{milestone_reporting_module_ui and milestone_reporting_module_server}
\usage{
milestone_reporting_module_ui(id)

milestone_reporting_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A named list. The list must contain a list named "tables".}

\item{config}{A named list.}
}
\description{
A shiny Module.
}
\keyword{internal}
