% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_submissions_module.R
\name{new_submissions_module_server}
\alias{new_submissions_module_server}
\title{New Submissions Module Server}
\usage{
new_submissions_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A named list. It must contain a list named "tables".}

\item{config}{A named list.}
}
\description{
New Submissions Module Server
}
