% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_submissions_module.R
\name{new_submissions_module_ui}
\alias{new_submissions_module_ui}
\title{New Submissions Module UI}
\usage{
new_submissions_module_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
New Submissions Module UI
}
