% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth_functions.R
\name{oauth_ui}
\alias{oauth_ui}
\title{OAUth UI}
\usage{
oauth_ui(request, ui_function)
}
\arguments{
\item{request}{Request Internal parameter for Shiny}

\item{ui_function}{A function that returns UI elements}
}
\description{
OAUth UI
}
