% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_timeline_module.R, R/plot_module.R
\name{file_upload_timeline_module_ui}
\alias{file_upload_timeline_module_ui}
\alias{plot_module_ui}
\alias{plot_module_server}
\title{File Upload Timeline Module UI}
\usage{
file_upload_timeline_module_ui(id)

plot_module_ui(id, title)

plot_module_server(id, data, config, plot_func, ...)
}
\arguments{
\item{id}{shiny id}

\item{title}{A string, the displayed title of the shinydashboard::box()}

\item{data}{A shiny::reactive() that returns a named list. The list must
contain a list named "tables".}

\item{config}{A shiny::reactive() that returns a named list.}

\item{plot_func}{A shiny::reactive() that returns a string that is the name of
a plotting function.}

\item{...}{Arguments to plotly::ggplotly}
}
\description{
A shiny Module.

A shiny Module.
}
\keyword{internal}
