% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_status_module.R
\name{publication_status_module_ui}
\alias{publication_status_module_ui}
\alias{publication_status_module_server}
\title{publication_status_module_ui and publication_status_module_server}
\usage{
publication_status_module_ui(id)

publication_status_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A named list. It must contain a list named "tables".}

\item{config}{A named list. It contain a list for each section:
\itemize{
\item publication_status
\item publication_disease
}}
}
\description{
A shiny Module.

A shiny Module.
}
\details{
Publication Status Module UI
}
