% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{recode_column_values}
\alias{recode_column_values}
\title{Recode Column Values
This function will recode a column of values in a tibble based o the list
passed. If no list is passed none of the values will be replaces unless
optional arguments are passed via ... to dplyr::recode such as .default
or .missing}
\usage{
recode_column_values(tbl, column, lst = NULL, ...)
}
\arguments{
\item{tbl}{A tibble

data <-  dplyr::tribble(
~col1, ~col2,
"a",   "c",
"a",   "d",
"b",   NA
)}

\item{column}{A string that is the name of a column in the tbl that will be
recoded}

\item{lst}{A named list where each name is a possible value in the column,
and the values are their replacements

lst <- list("a" = "x", "b" = "y", "c" = "z")}

\item{...}{Other arguments to dplyr::recode}
}
\description{
Recode Column Values
This function will recode a column of values in a tibble based o the list
passed. If no list is passed none of the values will be replaces unless
optional arguments are passed via ... to dplyr::recode such as .default
or .missing
}
