% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_column_values.R
\name{recode_column_with_config}
\alias{recode_column_with_config}
\title{Recode Column With Config}
\usage{
recode_column_with_config(tbl, config)
}
\arguments{
\item{tbl}{A Tibble}

\item{config}{A named list}
}
\description{
Recode Column With Config
}
