% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{replace_values_if_col_value_in_list}
\alias{replace_values_if_col_value_in_list}
\title{Replace Values If Column Values In List
This function replaces values in a list of columns, with a value, if a
given column has a supllied value.}
\usage{
replace_values_if_col_value_in_list(
  data,
  column,
  lst,
  columns,
  replace_value = NA_character_
)
}
\arguments{
\item{data}{A Tibble}

\item{column}{A string which is a name of a column in the data to check the
values in the lst}

\item{lst}{A list of values to check the column against}

\item{columns}{A list of strings that are column names.}

\item{replace_value}{A value to replace}
}
\description{
Replace Values If Column Values In List
This function replaces values in a list of columns, with a value, if a
given column has a supllied value.
}
