% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{safe_pluck_list}
\alias{safe_pluck_list}
\title{Safe Pluck List
This function is used to safely pluck named value from a list of named lists.
This will return a vector of values the length of the list. If the named list
doesn't have the name, an NA will be returned.}
\usage{
safe_pluck_list(lst, n)
}
\arguments{
\item{lst}{A list of named lists

list1 = list(
list("name" = "col1", "display_name" = "Column1", "type" = "x"),
list("name" = "col2", "type" = "x")
)}

\item{n}{A string. If the string is a name in the named lists, the value of
that string will be returned, otherwise NA.}
}
\description{
Safe Pluck List
This function is used to safely pluck named value from a list of named lists.
This will return a vector of values the length of the list. If the named list
doesn't have the name, an NA will be returned.
}
