% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dataframe_columns.R
\name{select_df_columns_with_config}
\alias{select_df_columns_with_config}
\title{Select Dataframe Columns With Config}
\usage{
select_df_columns_with_config(tbl, config)
}
\arguments{
\item{tbl}{A tibble

tbl <- dplyr::tribble(
~study, ~month,    ~year,
"s1",   "January", 2001L
)}

\item{config}{A list with a named list named "columns" that has an entry
for each column needed in the tibble. Each column must have a "name" field.

config <- list(
"columns" = list(
list("name" = "study"),
list("name" = "month")
)
)}
}
\description{
This function will select columns based on the Config.
Only columns in the Config will be selected.
}
