% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_snapshot_module.R
\name{summary_snapshot_module_ui}
\alias{summary_snapshot_module_ui}
\alias{summary_snapshot_module_server}
\title{summary_snapshot_module_server and summary_snapshot_module_server_ui}
\usage{
summary_snapshot_module_ui(id)

summary_snapshot_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}

\item{data}{A named list. The list must contain a list named "tables".}

\item{config}{A named list. The list must contain lists for each section:
\itemize{
\item header_text
\item overview_boxes
\item initiative_activity
\item resources_generated
\item file_upload_timeline
}}
}
\description{
A shiny Module.
}
\keyword{internal}
