% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_module.R
\name{synapse_module_server}
\alias{synapse_module_server}
\title{synapse_module_ui and synapse_module_server}
\usage{
synapse_module_server(id, syn, data_config)
}
\arguments{
\item{id}{A string, shiny id}

\item{syn}{A synapseclient$Synapse object that has been logged in}

\item{data_config}{A named list}
}
\description{
A shiny module.
}
\details{
Synapse Module Server
}
