% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_module2.R
\name{synapse_module_server2}
\alias{synapse_module_server2}
\title{synapse_module_ui and synapse_module_server}
\usage{
synapse_module_server2(id, syn, config)
}
\arguments{
\item{id}{A string, shiny id}

\item{syn}{A synapseclient$Synapse object that has been logged in}

\item{config}{a shiny::recative that returns a named list}
}
\description{
A shiny module.
}
\details{
Synapse Module Server2
}
