% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_module.R
\name{synapse_module_ui}
\alias{synapse_module_ui}
\title{synapse_module_ui and synapse_module_server}
\usage{
synapse_module_ui(id)
}
\arguments{
\item{id}{A string, shiny id}
}
\description{
A shiny module.
}
\details{
Synapse Module UI
}
