% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{concatenate_df_list_columns_with_config}
\alias{concatenate_df_list_columns_with_config}
\title{Concatenate Dataframe List Columns With Config
This function will concatenate list columns into character columns.
Any column of type "list:character" will be concatenated}
\usage{
concatenate_df_list_columns_with_config(data, config)
}
\arguments{
\item{data}{A Tibble

data <- dplyr::tribble(
~study,       ~month,
c("s1", "s2)  "January"
)}

\item{config}{A list with a named list named "columns" that has an entry
for each column needed in the tibble. Each column must have a "name", and
"type" field.

config <- list(
"columns" = list(
list(
"name" = "study",
"display_name" = "Consortium",
"type" = "list:character"
),
)
)}
}
\description{
Concatenate Dataframe List Columns With Config
This function will concatenate list columns into character columns.
Any column of type "list:character" will be concatenated
}
