% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_data_focus_plot}
\alias{create_data_focus_plot}
\title{Create Data Focus Plot}
\usage{
create_data_focus_plot(data, x, fill)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}
}
\description{
Create Data Focus Plot
}
