% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_publication_status_plot}
\alias{create_publication_status_plot}
\title{Create Publication Status Plot}
\usage{
create_publication_status_plot(
  data,
  x,
  fill,
  x_axis_text = NULL,
  y_axis_text = NULL
)
}
\arguments{
\item{data}{A Tibble}

\item{x}{A string that is the name of a column in the data}

\item{fill}{A string that is the name of a column in the data}

\item{x_axis_text}{A list of parameters to
axis.text.x = ggplot2::element_text()}

\item{y_axis_text}{A list of parameters to
axis.text.y = ggplot2::element_text()}
}
\description{
Create Publication Status Plot
}
