% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_functions.R
\name{format_plot_data_with_config}
\alias{format_plot_data_with_config}
\title{Format Plot Data With Config}
\usage{
format_plot_data_with_config(data, config)
}
\arguments{
\item{data}{A Tibble

data <- dplyr::tribble(
~consortium, ~year, ~month,
NA,          2000L, NA,
"c1",        20001, "January"
)}

\item{config}{A list with a named list named "columns" that has an entry
for each column needed in the tibble. Each column must have a "name",
and "type" field. Optional fields include "replace_values", "display_name",
"na_replace", and "deafult_replace".

config <- list(
"columns" = list(
list(
"name" = "consortium",
"display_name" = "Consortium",
"na_replace" = "Not Applicable",
"type" = "character"
),
)
)}
}
\description{
This function runs a tible through the main data cleanup functions befoire
being displayed in a plot or data table. See those functions for more
details.
}
