% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_summary_module.R
\name{study_summary_module_ui}
\alias{study_summary_module_ui}
\alias{study_summary_module_server}
\title{study_summary_module_ui and study_summary_module_server}
\usage{
study_summary_module_ui(id)

study_summary_module_server(id, data, config)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
